Saving player while he has <= 0 hp
closes #224

by @theblackboss
There is a bug.
if server is saving while player has <= 0 hp (but Player::onDie in Game::checkCreatures is not called yet cause of delay), the function player->preSave() will give him full hp and he wont die eventually (Player::onDie is not called, because the condition hp <= 0 is not fulfilled).
The solution is saving player with full hp and full mana (if hp <= 0), not giving him full hp.